Use --repository-dump-dir in Makefile
This intends to to standardize the the way we invoke the suite:
Makefile and GitHub action should, when possible use the same
options.

This does not try to wipe the previous dump contents if any exist:
this could lead to unexpected dump dir contents when tests are
modified and re-run: this does not affect the test suite and is solved
by deleting the directory and re-running tests.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>