Improve the "expected failures" experience
If we add more tests that we consider more or less optional features,
we need a way for clients to manage them.

xfails seems like it should work for that: Make it possible to
write the expected failures to a file next to the client-under-test
executable. This should make managing a longer xfail list reasonable.

This also means we don't need to repeat the xfails for go-tuf in two
places.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>