RepositorySimulator: Switch default keytype back to ecdsa
We made this rsa because the deterministic signatures allowed us to
workaround some metadata hashing tests. Now that we explicitly sign
everything, the workarounds are not needed.

Switch back to ecdsa: That seems to be more widely supported (RSA
support is patchy because of the historic flip-flopping the spec did
between PSS and pkcs1v15).

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>